+++ R/run_ICA.R
@@ -18,12 +18,14 @@ run_ICA <- function(data, ...) {
 #' @param method "sobi" (default), "fastica", or "infomax".
 #' @param maxit Maximum number of iterations of the Infomax and Fastica ICA
 #'   algorithms.
+#' @param tol Convergence tolerance for fastica and infomax. Defaults to 1e-06.
 #' @describeIn run_ICA Run ICA on an \code{eeg_epochs} object
 #' @export
 
 run_ICA.eeg_epochs <- function(data,
                                method = "sobi",
                                maxit = 500,
+                               tol = 1e-6,
                                ...) {
 
   rank_check <- Matrix::rankMatrix(as.matrix(data$signals))
@@ -34,7 +36,12 @@ run_ICA.eeg_epochs <- function(data,
   #}
 
   if (method == "sobi") {
-    ica_obj <- sobi_ICA(data)
+    if (!requireNamespace("JADE", quietly = TRUE)) {
+      stop("Package \"JADE\" needed to use SOBI. Please install it.",
+           call. = FALSE)
+    }
+    ica_obj <- sobi_ICA(data,
+                        maxiter = maxit)
   } else {
 
     if (!requireNamespace("ica", quietly = TRUE)) {
@@ -49,12 +56,14 @@ run_ICA.eeg_epochs <- function(data,
     if (method == "fastica") {
       ICA_out <- ica::icafast(data$signals,
                               rank_check,
-                              maxit = maxit)
+                              maxit = maxit,
+                              tol = tol)
     } else if (method == "infomax") {
       ICA_out <- ica::icaimax(data$signals,
                               rank_check,
                               maxit = maxit,
-                              fun = "ext")
+                              fun = "ext",
+                              tol = tol)
     } else {
       stop("Unknown method; available methods are sobi, fastica, and infomax.")
     }
@@ -75,7 +84,8 @@ run_ICA.eeg_epochs <- function(data,
                        events = data$events,
                        chan_info = data$chan_info,
                        srate = data$srate,
-                       epochs = data$epochs)
+                       epochs = data$epochs,
+                       algorithm = method)
     }
   ica_obj
 }
@@ -85,11 +95,13 @@ run_ICA.eeg_epochs <- function(data,
 #' Internal function for running SOBI ICA on an \code{eeg_epochs} object
 #'
 #' @param data Data to be ICAed.
+#' @param maxiter Maximum number of iterations of the joint diagonalization
 #' @author A. Belouchrani and A. Cichocki. Adapted to R by Matt Craddock
 #'   \email{matt@@mattcraddock.com}
 #' @keywords internal
 
-sobi_ICA <- function(data) {
+sobi_ICA <- function(data,
+                     maxiter) {
 
   n_epochs <- length(unique(data$timings$epoch))
   n_channels <- ncol(data$signals)
@@ -104,17 +116,21 @@ sobi_ICA <- function(data) {
   #do by epochs
   amp_matrix <- split(data$signals, data$timings$epoch)
   amp_matrix <- lapply(amp_matrix,
-                       function(x) sweep(x,
-                                         2,
-                                         Matrix::colMeans(x))
-                       )
-  amp_matrix <- as.matrix(do.call("rbind", amp_matrix))
+                         function(x) sweep(x,
+                                           2,
+                                           Matrix::colMeans(x))
+                         )
+  # # amp_matrix <- lapply(amp_matrix,
+  # #                      function(x) scale(x, scale = FALSE))
+   amp_matrix <- as.matrix(do.call("rbind", amp_matrix))
+  #amp_matrix <- sweep(data$signals, 2, Matrix::colMeans(data$signals))
   SVD_amp <- svd(amp_matrix)
 
   ## get the psuedo-inverse of the diagonal matrix, multiply by singular
   ## vectors
-  Q <- MASS::ginv(diag(SVD_amp$d), tol = 0) %*% t(SVD_amp$v) # whitening matrix
-  amp_matrix <- Q %*% t(amp_matrix)
+   Q <- MASS::ginv(diag(SVD_amp$d), tol = 0) %*% t(SVD_amp$v) # whitening matrix
+   amp_matrix <- Q %*% t(amp_matrix)
+  #amp_matrix <- amp_matrix %*% SVD_amp$v
 
   ## reshape to reflect epoching structure
   dim(amp_matrix) <- c(n_channels,
@@ -142,18 +158,25 @@ sobi_ICA <- function(data) {
     Rxp <- Reduce("+",
                   Rxp)
     Rxp <- Rxp / n_epochs
-    M[, u:(u + n_channels - 1)] <- norm(Rxp, "F") * (Rxp) #  Frobenius norm
+    M[, u:(u + n_channels - 1)] <- norm(Rxp, "F") * (Rxp)
   }
 
   epsil <- 1 / sqrt(N) / 100
-  V <- joint_diag(M,
-                  epsil,
-                  n_channels,
-                  pm) # V = sphered unmixing matrix
+  V <- JADE::rjd(t(M),
+                 eps = epsil,
+                 maxiter = maxiter)$V
 
   ## create mixing matrix for output
   mixing_matrix <- MASS::ginv(Q, tol = 0) %*% V
-  unmixing_matrix <- t(V)
+
+  unmixing_matrix <- MASS::ginv(mixing_matrix, tol = 0)
+  # rescale vecs
+  scaling <- sqrt(colMeans(mixing_matrix^2))
+
+  unmixing_matrix <- sweep(unmixing_matrix, MARGIN = 1, scaling, `*`) # scaled weights
+
+  mixing_matrix <- MASS::ginv(unmixing_matrix %*% diag(ncol(unmixing_matrix)),
+                              tol = 0)
 
   dim(amp_matrix) <- c(n_channels,
                        n_times * n_epochs)
@@ -176,76 +199,10 @@ sobi_ICA <- function(data) {
                      events = data$events,
                      chan_info = data$chan_info,
                      srate = data$srate,
-                     epochs = data$epochs)
+                     epochs = data$epochs,
+                     algorithm = "sobi")
   ica_obj
 }
-
-#' Joint diagonalization for SOBI
-#'
-#' Drop-in replacement for JADE::rjd, which often fails to converge
-#'
-#' @param M Numeric matrix
-#' @param eps Convergence tolerance
-#' @param pm number of channels * number of timepoints
-#' @param n_channels number of channels
-#' @author A. Belouchrani and A. Cichocki. Adapted to R by Matt Craddock
-#'   \email{matt@@mattcraddock.com}
-#' @keywords internal
-
-joint_diag <- function(M,
-                       eps,
-                       n_channels,
-                       pm) {
-
-  epsil <- eps
-  continue <- TRUE
-  V <- diag(n_channels)
-  step_n <- 0
-
-  while (isTRUE(continue)) {
-    continue <- FALSE
-    for (p in 1:(n_channels - 1)) {
-      for (q in ((p + 1):n_channels)) {
-
-        P_seq <- seq(p, pm, n_channels)
-        Q_seq <- seq(q, pm, n_channels)
-
-        # Perform Givens rotation
-        g <- rbind(M[p, P_seq] - M[q, Q_seq],
-                   M[p, Q_seq] + M[q, P_seq],
-                   1i * (M[q, P_seq] - M[p, Q_seq]))
-        eigs <- eigen(Re(tcrossprod(g))) # %*% t(g)))
-        sort_eigs <- sort(eigs$`values`, index.return = T)
-        angles <- eigs$vectors[, sort_eigs$ix[3]]
-        angles <- sign(angles[1]) * angles
-        c_r <- sqrt(0.5 + angles[1] / 2)
-        sr <- Re(0.5 * (angles[2] - 1i * angles[3]) / c_r)
-        sc <- Conj(sr)
-        conv_check <- abs(sr) > epsil
-        continue <- (continue | conv_check)
-        if (conv_check) {# Update the M and V matrices
-          colp <- M[, P_seq]
-          colq <- M[, Q_seq]
-          M[, P_seq] <- c_r * colp + sr * colq
-          M[, Q_seq] <- c_r * colq - sc * colp
-          rowp <- M[p, ]
-          rowq <- M[q, ]
-          M[p, ] <- c_r * rowp + sc * rowq
-          M[q, ] <- c_r * rowq - sr * rowp
-          temp <- V[, p]
-          V[, p] <- c_r * V[, p] + sr * V[, q]
-          V[, q] <- c_r * V[, q] - sc * temp
-        }
-      }
-    }
-    step_n <- step_n + 1
-    message("Iteration = ", step_n, "\n")
-  }
-  V
-}
-
-
-#
 
 apply_ica <- function(data, ...) {
   UseMethod("apply_ica", data)
@@ -260,12 +217,12 @@ apply_ica.eeg_ICA <- function(data, comps, ...) {
   names(new_dat) <- data$chan_info$electrode
   out <-
     eeg_data(new_dat,
-           data$srate,
-           data$events,
-           data$chan_info,
-           data$timings,
-           data$continuous,
-           reference = NULL
+             data$srate,
+             data$events,
+             data$chan_info,
+             data$timings,
+             data$continuous,
+             reference = NULL
           )
   class(out) <- c("eeg_epochs", "eeg_data")
   out
@@ -279,7 +236,7 @@ apply_ica.eeg_epochs <- function(data, decomp, comps, ...) {
     new_mixmat <- decomp$mixing_matrix[1:(ncomps)]
     new_mixmat[, comps] <- 0
     source_sigs <- as.matrix(data$signals) %*%
-      as.matrix(decomp$unmixing_matrix[, 1:ncomps])
+      t(as.matrix(decomp$unmixing_matrix[, 1:ncomps]))
 
     new_dat <- as.matrix(new_mixmat) %*% t(source_sigs)
     new_dat <- as.data.frame(t(new_dat))